update agp dep and some fix

+++ core/data/src/main/java/com/naaammme/bbspace/core/data/repository/VodDanmakuRepoImpl.kt
@@ -65,7 +65,7 @@ class VodDanmakuRepoImpl @Inject constructor(
                 progressMs = elem.progress,
                 mode = elem.mode,
                 fontSize = elem.fontsize,
-                color = elem.color.toInt(),
+                color = elem.color,
                 midHash = elem.midHash,
                 content = elem.content,
                 createdAtEpochSecond = elem.ctime,
